fix sys.argv for fileinput
An earlier patch of mine broke the fileinput behavior.
This patch fixes the problem, which was twofold.

First, the argparse object was constructed improperly.  With
nargs='*', an action of 'store' (the default) is preferable; the
previous value 'append' means that the result will be a one-element
list holding a list of results.

Second, fileinput ignores sys.argv[0], so we must be careful to only
store to the appropriate slice.